refactor session to allow multiple tabs at once

+++ app/routes.py
@@ -3,35 +3,32 @@ from flask import render_template, url_for, request, session, send_from_director
 import create_rim_weight as engine
 import quantipy as qp 
 import pandas as pd
-import os, sys, traceback
+import os, sys, traceback, uuid
 import pyreadstat
 
 @app.route('/')
 @app.route('/index')
 def index():
     # load main page
-    # reset and initialize session
-    session.clear()
-    session['active'] = None
     return render_template('index.html')
 
 @app.route('/get-meta', methods=['POST'])
 def process_file():
     try:         
-        session['active'] = True
-        print(request.files)
+        uid = str(uuid.uuid4())  # use uid to track multiple instances in separate tabs/windows
+        # reset and initialize session
+        session[uid] = {}
+        session[uid]['active'] = True
         file = request.files['file']  # .sav file user selected
         filename = file.filename
-        print("name of file:")
-        print(filename)
-        session['filename'] = filename
+        print("name of file: " + filename)
+        session[uid]['filename'] = filename
         # create temp directory if it does not exist
         if not os.path.isdir('./temp'):
             print('Creating temp directory')
             os.system(f"mkdir {'./temp'}")
         file.save('./temp/%s' %(filename))
         print("Temporary file saved")
-        print(session)
 
 
         # read dataset from saved .sav file
@@ -46,7 +43,7 @@ def process_file():
         if ('uuid' not in ds.meta()['columns']):
             return {'success': 'false', 'message': 'This file does not have a uuid variable.'}
 
-        return {'success': 'true', 'meta_data_array': meta_data, 'meta_data_obj': ds.meta()['columns']}
+        return {'success': 'true', 'uid': uid, 'meta_data_array': meta_data, 'meta_data_obj': ds.meta()['columns']}
 
     except Exception as e:
         print(e)
@@ -58,20 +55,17 @@ def process_file():
 @app.route('/compute-weights', methods=['POST'])
 def compute_weights():
     req_data = request.json
-    # print(req_data)
+    uid = req_data['uid']
     target_variables = req_data['targetVariables']
     target_mapping = req_data['targetMapping']
     grouping_variable = req_data['groupingVariable']
     weight_name = req_data['weightName']
 
-    print(weight_name)
-
-    print(session)
-    file_name = session['filename'] ## make special error if this doesn't exist
+    file_name = session[uid]['filename'] ## make special error if this doesn't exist
     try:        
         file_location, syntax_location, crosstabs, report = engine.weight_data(target_variables, target_mapping, grouping_variable, file_name, weight_name=weight_name)
-        session['weighted_location'] = file_location
-        session['syntax_location'] = syntax_location
+        session[uid]['weighted_location'] = file_location
+        session[uid]['syntax_location'] = syntax_location
         return {'success': 'true', 'location': file_location, 'syntax': syntax_location, 'crosstabs': crosstabs, 'report': report}
     except Exception as e:
         print(type(e))
@@ -85,30 +79,37 @@ def download(filename):
     uploads = os.path.join(os.path.dirname(app.root_path), 'temp')
     return send_from_directory(directory=uploads, filename=filename, as_attachment=True)
 
-@app.route('/close', methods=['GET', 'POST'])
+@app.route('/close', methods=['GET','POST'])
 def close_resources():
     print('Closing resources')
-    # delete sav files
-    original_file_name = session.get('filename')
-    weighted_file_name = session.get('weighted_location')
-    syntax_file_name = session.get('syntax_location')
-    if original_file_name:
-        try:
-            os.remove('./temp/' + original_file_name)
-        except FileNotFoundError:
-            print("This file does not exist")
-    if weighted_file_name:
-        try:
-            os.remove(weighted_file_name)
-        except FileNotFoundError:
-            print("This weighted file is no longer in the folder")
-    if syntax_file_name:
-        try:
-            os.remove(syntax_file_name)
-        except FileNotFoundError:
-            print("This syntax file is no longer in the folder")
+    uid = ''
+    if request.args: 
+        uid = request.args.get('uid')
 
-    # destroy session
-    session.clear()
+    if uid:    
+        original_file_name = session[uid].get('filename')
+        weighted_file_name = session[uid].get('weighted_location')
+        syntax_file_name = session[uid].get('syntax_location')
+        
+        # delete spss files
+        if original_file_name:
+            try:
+                os.remove('./temp/' + original_file_name)
+            except FileNotFoundError:
+                print("This file does not exist")
+        if weighted_file_name:
+            try:
+                os.remove(weighted_file_name)
+            except FileNotFoundError:
+                print("This weighted file is no longer in the folder")
+        if syntax_file_name:
+            try:
+                os.remove(syntax_file_name)
+            except FileNotFoundError:
+                print("This syntax file is no longer in the folder")
+
+        # destroy session
+        session[uid] = {}
+    
 
     return {'success': 'true'}
